Finished reworking topology panel
