--- Net2Plan-GUI/Net2Plan-GUI-Plugins/Net2Plan-NetworkDesign/src/main/java/com/net2plan/gui/plugins/networkDesign/topologyPane/IconCreator.java
@@ -1,70 +1,0 @@
-package com.net2plan.gui.plugins.networkDesign.topologyPane;
-
-import com.net2plan.utils.Pair;
-
-import javax.imageio.ImageIO;
-import javax.swing.*;
-import java.awt.*;
-import java.awt.image.BufferedImage;
-import java.io.File;
-
-public class IconCreator {
-
-	public static void main(String[] args) throws Exception 
-	{
-//		Pair<Image,Shape> pair = new IconCreator ().getImage ("src/main/resources/resources/icons/router.jpg" , 50);
-		Pair<Image,Shape> pair = new IconCreator ().getImage ("src/main/resources/resources/icons/router.png" , 50);
-//		Pair<Image,Shape> pair = new IconCreator ().getImage ("src/main/resources/resources/icons/googlelogo_color_120x44dp.png" , 50);
-
-		//BufferedImage img = ImageIO.read(new File("src/main/resources/resources/icons/googlelogo_color_120x44dp.png")); 
-		BufferedImage img = ImageIO.read(new File("src/main/resources/resources/icons/imagen1.png"));
-		
-		
-		
-		
-		JFrame frame = new JFrame("Example");
-		frame.getContentPane().setLayout(new FlowLayout());
-		frame.getContentPane().add(new JLabel(new ImageIcon(img)));
-//		frame.getContentPane().add(new JLabel(new ImageIcon(img2)));
-//		frame.getContentPane().add(new JLabel(new ImageIcon(img3)));
-		frame.pack();
-		frame.setVisible(true);JPanel panel = new ImagePanel(img);
-//		panel.repaint();
-//	    frame.add(panel);
-//	    frame.pack();
-//	    frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-//	    frame.setVisible(true);   
-//		// get Image
-
-	}
-	
-	public Pair<Image,Shape> getImage (String file , int heightPixels) throws Exception
-	{
-		System.out.println(new File (file).exists());
-		System.out.println(new File (file).getAbsolutePath());
-		BufferedImage img = ImageIO.read(new File(file)); 
-		System.out.println("here");
-        return Pair.of(img,null);
-	}
-
-	public static class ImagePanel extends JPanel 
-	{
-
-		BufferedImage image;
-		Dimension size;
-
-		public ImagePanel(BufferedImage image) 
-		{
-		    this.image = image;
-		    this.size = new Dimension();
-		    System.out.println(image.getWidth() + " , " + image.getHeight());
-		    size.setSize(image.getWidth(), image.getHeight());
-//		    this.setBackground(Color.WHITE);
-//		    this.setBorder(BorderFactory.createLineBorder(Color.RED, 1));
-		}
-		@Override
-		public Dimension getPreferredSize() { return size; }
-
-	}
-
-}
