+++ Net2Plan-GUI/Net2Plan-GUI-Plugins/Net2Plan-NetworkDesign/src/main/java/com/net2plan/gui/plugins/networkDesign/topologyPane/TopologyTopBar.java
@@ -218,7 +218,7 @@ public class TopologyTopBar extends JToolBar implements StateObserver, ActionLis
             canvas.refresh();
         } else if (src == btn_takeSnapshot)
         {
-            topologyPanel.takeSnapshot();
+            canvas.takeSnapshot();
         } else if (src == btn_zoomIn)
         {
             canvas.zoomIn();
